Refactor of duplicate code.
